TDA Results

+++ jupyter/04_cubical_persistence.py
@@ -32,23 +32,26 @@ for tidx in range(len(transcriptomes)):
     if not os.path.isdir(tdst):
         os.mkdir(tdst)
         
-    cidx = 6
     for cidx in range(len(tiffs)):
-        img = tf.imread(tiffs[cidx])
-        if max(img.shape) > maximgsize:
-            zoom = maximgsize/max(img.shape)
-            print(tiffs[cidx], '\nResized', img.shape, 'by a factor of ', zoom)
-            img = ndimage.zoom(img, zoom = zoom, order=1, mode='reflect')
-            print('Now', img.shape, '\n----')
-
-        cc = gd.CubicalComplex(top_dimensional_cells = img)
-        pers = cc.persistence(homology_coeff_field=2, min_persistence=1)
+        
         filename = tdst + 'sublevel_' + os.path.splitext(os.path.split(tiffs[cidx])[1])[0] + '.json'
-        with open(filename, 'w') as f:
-            json.dump(pers,f)
+        
+        if not os.path.isfile(filename):        
+            img = tf.imread(tiffs[cidx])
+            if max(img.shape) > maximgsize:
+                zoom = maximgsize/max(img.shape)
+                print(tiffs[cidx], '\nResized', img.shape, 'by a factor of ', zoom)
+                img = ndimage.zoom(img, zoom = zoom, order=1, mode='reflect')
+                print('Now', img.shape, '\n----')
 
-        cc = gd.CubicalComplex(top_dimensional_cells = 255 - img)
-        pers = cc.persistence(homology_coeff_field=2, min_persistence=1)
-        filename = tdst + 'superlevel_' + os.path.splitext(os.path.split(tiffs[cidx])[1])[0] + '.json'
-        with open(filename, 'w') as f:
-            json.dump(pers,f)
+            cc = gd.CubicalComplex(top_dimensional_cells = img)
+            pers = cc.persistence(homology_coeff_field=2, min_persistence=1)
+            filename = tdst + 'sublevel_' + os.path.splitext(os.path.split(tiffs[cidx])[1])[0] + '.json'
+            with open(filename, 'w') as f:
+                json.dump(pers,f)
+
+            cc = gd.CubicalComplex(top_dimensional_cells = 255 - img)
+            pers = cc.persistence(homology_coeff_field=2, min_persistence=1)
+            filename = tdst + 'superlevel_' + os.path.splitext(os.path.split(tiffs[cidx])[1])[0] + '.json'
+            with open(filename, 'w') as f:
+                json.dump(pers,f)
